Rearranging activation macros
